move /dashboard contents into /insights + renaming work
