--- pontoon/dashboard/templates/dashboard/dashboard.html
@@ -1,116 +1,0 @@
-{% extends "base.html" %}
-{% import 'widgets/heading_info.html' as HeadingInfo %}
-{% import 'dashboard/widgets/locale_list.html' as LocaleList %}
-
-{% block title %}Community Health Dashboard{% endblock %}
-
-{% block class %}teams{% endblock %}
-
-{% block heading %}
-  <section id="heading">
-    <div class="container clearfix">
-      <h1>
-        {{
-          HeadingInfo.heading_item(
-          title='Hey, ' ~ user.name_or_email ~ '!',
-          link=url('pontoon.dashboard'))
-        }}
-      </h1>
-
-      <ul class="details">
-        {{
-          HeadingInfo.details_item(
-          title='Number of teams',
-          class='teams-count',
-          value=locales|length)
-        }}
-        {% if top_instances %}
-          {{
-            HeadingInfo.details_item(
-            title='Most translations',
-            class='most-translations',
-            value=top_instances['most_translations'],
-            value_link=url('pontoon.teams.team', top_instances['most_translations'].code))
-          }}
-          {{
-            HeadingInfo.details_item(
-            title='Most unreviewed',
-            class='most-suggestions',
-            value=top_instances['most_suggestions'],
-            value_link=url('pontoon.teams.team', top_instances['most_suggestions'].code))
-          }}
-          {{
-            HeadingInfo.details_item(
-            title='Most missing strings',
-            class='missing-strings',
-            value=top_instances['most_missing'],
-            value_link=url('pontoon.teams.team', top_instances['most_missing'].code))
-          }}
-          {{
-            HeadingInfo.details_item(
-            title='Most enabled strings',
-            class='most-strings',
-            value=top_instances['most_strings'],
-            value_link=url('pontoon.teams.team', top_instances['most_strings'].code))
-          }}
-        {% endif %}
-      </ul>
-
-      {{ HeadingInfo.progress_chart() }}
-      {{ HeadingInfo.progress_chart_legend(all_locales_stats) }}
-    </div>
-  </section>
-{% endblock %}
-
-{% block bottom %}
-  <section id="main" class="clearfix">
-    <div class="container">
-      <menu class="controls clearfix">
-        <div class="search-wrapper small clearfix">
-          <div class="icon fas fa-search"></div>
-          <input
-            class="table-filter"
-            type="search"
-            autocomplete="off"
-            autofocus
-            placeholder="Filter teams"
-            data-filter="td:nth-child(1)"
-          />
-        </div>
-        <div class="button-container">
-          <a
-            class="button"
-            id="edit-config"
-            href="{{ url('pontoon.dashboard.config') }}"
-            >Edit Configuration
-          </a>
-          <button class="button" id="view-toggle">Show scores</button>
-        </div>
-      </menu>
-
-      {{ LocaleList.header() }}
-
-      {% for locale in locales %}
-        {% set main_link = url('pontoon.teams.team', locale.code) %}
-
-        {{ LocaleList.item(locale, main_link, curr_snapshot=current_snapshots.get(locale.id), base_deltas=snapshot_base_deltas.get(locale.id), score_deltas=snapshot_score_deltas.get(locale.id), columns=columns) }}
-      {% endfor %}
-
-      {{ LocaleList.footer(request=True, form=form) }}
-
-      {% if not locales %}
-        <div class="no-locales">
-          <p>No locales selected.</p>
-        </div>
-      {% endif %}
-    </div>
-  </section>
-{% endblock %}
-
-{% block extend_css %}
-  {% stylesheet 'dashboard' %}
-{% endblock %}
-
-{% block extend_js %}
-  {% javascript 'dashboard' %}
-{% endblock %}
