+++ pontoon/insights/static/css/config.css
