+++ pontoon/insights/static/js/insights.js
@@ -9,6 +9,28 @@ const longMonthFormat = new Intl.DateTimeFormat('en', {
 
 const style = getComputedStyle(document.body);
 
+$('body').on('click', '#view-toggle', function (e) {
+  e.stopPropagation();
+
+  const table = $('.locale-list');
+  table.toggleClass('show-score-view');
+
+  const showScores = table.hasClass('show-score-view');
+
+  // Keep each cells sort key in sync
+  table.find('td.cell').each(function () {
+    const td = $(this);
+    const key = showScores
+      ? td.attr('data-score-sort')
+      : td.attr('data-base-sort');
+    if (key !== undefined) {
+      td.attr('data-sort', key);
+    }
+  });
+
+  $('#view-toggle').text(showScores ? 'Show default' : 'Show scores');
+});
+
 // eslint-disable-next-line no-var
 var Pontoon = (function (my) {
   return $.extend(true, my, {
