+++ pontoon/insights/urls.py
@@ -1,13 +1,19 @@
 from django.urls import path
 
-from . import views
+from pontoon.insights.views import insights, insights_config
 
 
 urlpatterns = [
     # Insights page
     path(
         "insights/",
-        views.insights,
+        insights,
         name="pontoon.insights",
+    ),
+    # Insights config page
+    path(
+        "insights/config",
+        insights_config,
+        name="pontoon.insights.config",
     ),
 ]
