+++ pontoon/urls.py
@@ -79,7 +79,6 @@ urlpatterns = [
     path("", include("pontoon.tour.urls")),
     path("", include("pontoon.tags.urls")),
     path("", include("pontoon.search.urls")),
-    path("", include("pontoon.dashboard.urls")),
     path("", include("pontoon.sync.urls")),
     path("", include("pontoon.projects.urls")),
     path("", include("pontoon.machinery.urls")),
